Move test store helper
