+++ tests/state/Data/Components/actions/addComponentsFromView-test.js
@@ -1,6 +1,6 @@
 import {assert} from 'chai';
 import actions from '../../../../../src/state/Data/Components/actions';
-import getStoreSet from '../../../_common/helpers/store';
+import getStoreSet from '../../../../store';
 
 describe('state/Data/Components/actions/addComponentsFromView', function () {
 	it('Dispatch addComponentsFromView action', function () {
