+++ tests/state/Data/Components/actions/componentUseClear-test.js
@@ -1,6 +1,6 @@
 import {assert} from 'chai';
 import actions from '../../../../../src/state/Data/Components/actions';
-import getStoreSet from '../../../_common/helpers/store';
+import getStoreSet from '../../../../store';
 
 describe('state/Data/Components/actions/componentUseClear', function () {
 	it('Dispatch clearComponent when is component registered in state', function () {
