+++ tests/state/Data/Components/actions/ensureWithFilterByActive-test.js
@@ -5,7 +5,7 @@ import _ from 'lodash';
 import slash from 'slash';
 import actions from '../../../../../src/state/Data/Components/actions';
 import {resetFetch, setFetch} from '../../../../../src/state/_common/request';
-import getStoreSet from '../../../_common/helpers/store';
+import getStoreSet from '../../../../store';
 
 import AttributeDataReducer from '../../../../../src/state/Data/AttributeData/reducers';
 import AttributeRelationsReducer from '../../../../../src/state/Data/AttributeRelations/reducers';
