+++ tests/state/Data/SpatialDataSources/actions/add-test.js
@@ -1,6 +1,6 @@
 import {assert} from 'chai';
 import actions from '../../../../../src/state/Data/AttributeDataSources/actions';
-import getStoreSet from '../../../_common/helpers/store';
+import getStoreSet from '../../../../store';
 describe('state/Data/SpatialDataSources/actions/add', function () {
 	it('Dispatch add action', function () {
 		const storeHelpers = getStoreSet();
