+++ tests/state/Data/actions/loadMissingAttributeData-test.js
@@ -12,7 +12,7 @@ import SpatialDataReducer from '../../../../src/state/Data/SpatialData/reducers'
 import SpatialRelationsReducer from '../../../../src/state/Data/SpatialRelations/reducers';
 import SpatialDataSourcesReducer from '../../../../src/state/Data/SpatialDataSources/reducers';
 import AppReducers from '../../../../src/state/App/reducers';
-import getStoreSet from '../../_common/helpers/store';
+import getStoreSet from '../../../store';
 import {
 	responseWithRelationsSpatialAndAttributeData_1,
 	responseWithRelationsSpatialAndAttributeData_2,
