+++ tests/state/Maps/actions/layerUse-test.js
@@ -4,7 +4,7 @@ import {isEqual as _isEqual} from 'lodash';
 import {createStore, combineReducers} from 'redux';
 import {setState} from '@jvitela/recompute';
 import actions from '../../../../src/state/Maps/actions';
-import getStoreSet from '../../_common/helpers/store';
+import getStoreSet from '../../../store';
 
 import {resetFetch, setFetch} from '../../../../src/state/_common/request';
 import MapsReducers from '../../../../src/state/Maps/reducers';
