+++ tests/state/_common/actions/loadKeysPage-test.js
@@ -28,13 +28,6 @@ const tests = [
 				},
 			},
 		}),
-		getOptions: params => {
-			return {
-				getSubstate: params.getSubstate,
-				dataType: params.dataType,
-				categoryPath: params.categoryPath,
-			};
-		},
 		setFetch: (dataType, categoryPath) => (url, options) => {
 			assert.strictEqual(
 				`http://localhost/rest/${categoryPath}/filtered/${dataType}`,
@@ -78,6 +71,9 @@ const tests = [
 
 const dataType = 'testStore';
 const categoryPath = 'metadata';
-describe('loadKeysPage', testBatchRunner(dataType, categoryPath, tests, commonActions, actionTypes));
+describe(
+	'loadKeysPage',
+	testBatchRunner(dataType, categoryPath, tests, commonActions, actionTypes)
+);
 
 export default tests;
